Refactored DatabaseConnection to SqliteConnection to better reflect to type of database.
