--- src/main/java/com/example/napolinights/model/DatabaseConnection.java
@@ -1,25 +1,0 @@
-package com.example.napolinights.model;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class DatabaseConnection {
-    private static Connection instance = null;
-
-    private DatabaseConnection() {
-        String url = "jdbc:sqlite:database.db";
-        try {
-            instance = DriverManager.getConnection(url);
-        } catch (SQLException sqlEx) {
-            System.err.println(sqlEx);
-        }
-    }
-
-    public static Connection getInstance() {
-        if (instance == null) {
-            new DatabaseConnection();
-        }
-        return instance;
-    }
-}
